Added dash
Player can now dash with the space bar.
