+++ GameSim_UnityProject/ProjectDev_UnityProject/Assets/_Prefabs/Dummy Player.prefab
@@ -54,15 +54,15 @@ Rigidbody2D:
   m_Simulated: 1
   m_UseFullKinematicContacts: 0
   m_UseAutoMass: 0
-  m_Mass: 1
-  m_LinearDrag: 0
-  m_AngularDrag: 0.05
+  m_Mass: 1.75
+  m_LinearDrag: 6.75
+  m_AngularDrag: 4.75
   m_GravityScale: 0
   m_Material: {fileID: 0}
   m_Interpolate: 0
   m_SleepingMode: 1
   m_CollisionDetection: 0
-  m_Constraints: 0
+  m_Constraints: 4
 --- !u!61 &61015909524525122
 BoxCollider2D:
   m_ObjectHideFlags: 1
@@ -99,7 +99,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 1d9a2b79a61772c4bb436c8428e4a25b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  speed: 5
+  speed: 40
+  dashSpeed: 800
+  timeBetweenDashes: 2
 --- !u!212 &212323944088977610
 SpriteRenderer:
   m_ObjectHideFlags: 1
